+++ .gitignore
@@ -64,7 +64,7 @@ replay_pid*
 
 
 # dependencies
-/node_modules
+project-front/node_modules
 /.pnp
 .pnp.js
 
@@ -88,3 +88,26 @@ yarn-error.log*
 /idea
 
 
+# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
+
+dist
+dist-ssr
+*.local
+
+# Editor directories and files
+.vscode/*
+!.vscode/extensions.json
+.idea
+.DS_Store
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
